更新 v1.4
## [v1.4] - 2021-7-24
* 1.更新lvgl
* 2.更新MapConv，统一使用GPS_Transform校准坐标
* 3.更新HAL::GPS模拟，支持航向和速度模拟
* 4.修改HAL::GPS_Info_t的 compass -> course
* 5.使用合入lvgl主线的lv_anim_timeline，使用lv_anim_timeline_wrapper封装动画组
* 6.更新StartUp页面，新的开机动画
* 7.LV_IMG_CACHE_DEF_SIZE设置为0

+++ Software/X-Track/USER/App/Utils/TrackFilter/TrackPointFilter.cpp
@@ -283,7 +283,7 @@ double TrackPointFilter::QuickSort(double num)
     long i;
     float x2, y;
     const float threehalfs = 1.5f;
-    x2 = num * 0.5f;
+    x2 = (float)num * 0.5f;
     y = (float)num;
     i = *(long*)&y;
     i = 0x5f3759df - (i >> 1);
